完成redis实现抢红包后台基本逻辑，测试了使用redis和不使用redis直接cas操作数据库的性能差别，后续会完善整个ssm+redis项目,添加新的前端界面与后台功能。

+++ src/main/java/com/lujunqiu/service/UserRedPacketService.java
@@ -29,5 +29,5 @@ public interface UserRedPacketService {
      * 1-成功抢到红包，且不是最后一个红包
      * 2-成功抢到红包，且是最后一个红包
      */
-    int grapRedPacketByRedis(int redPacketId, int userId);
+    long grapRedPacketByRedis(int redPacketId, int userId);
 }
